pyproject.toml
[project]
name = "github-commit-reviewer"
version = "0.1.0"
description = "A Python tool that analyzes GitHub commits using GPT-4o"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "PyGithub>=2.7.0",
    "openai>=1.98.0",
    "python-dotenv>=1.1.1",
]

[project.scripts]
github-commit-reviewer = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
exclude = [
    ".venv",
    "__pycache__",
    "*.pyc",
    ".env",
    "commits_*.md",
    "*_review.json",
]
